Restructure parseSty into a separate file and add testing for it.
