--- lib/npm-debug.log
@@ -1,18 +1,0 @@
-0 info it worked if it ends with ok
-1 verbose cli [ 'node', '/usr/bin/npm', 'test' ]
-2 info using npm@1.3.6
-3 info using node@v0.10.42
-4 error Error: ENOENT, open '/home/jrp4/Documents/i_just_want_to_avoid_directory_conflicts/texvcjs/lib/package.json'
-5 error If you need help, you may report this log at:
-5 error     <http://github.com/isaacs/npm/issues>
-5 error or email it to:
-5 error     <npm-@googlegroups.com>
-6 error System Linux 4.6.4-201.fc23.x86_64
-7 error command "node" "/usr/bin/npm" "test"
-8 error cwd /home/jrp4/Documents/i_just_want_to_avoid_directory_conflicts/texvcjs/lib
-9 error node -v v0.10.42
-10 error npm -v 1.3.6
-11 error path /home/jrp4/Documents/i_just_want_to_avoid_directory_conflicts/texvcjs/lib/package.json
-12 error code ENOENT
-13 error errno 34
-14 verbose exit [ 34, true ]
