+++ lib/texutil.js
@@ -3,6 +3,8 @@
 // every time that parse() is called.
 "use strict";
 
+var parseSty = require("./parseSty");
+
 // track all known function names, so we can give good errors for unknown
 // functions.
 var all_functions = module.exports.all_functions = Object.create(null);
@@ -42,128 +44,20 @@ module.exports.latex_function_names = arr2set([
     "\\sin", "\\sinh", "\\sup", "\\tan", "\\tanh"
 ]);
 
-function findBracketSections(string) {
-    var sections = [[""]];
-    var char;
-    var level = 0;
-    for (var i = 0; i < string.length; i++) {
-        char = string.charAt(i);
-        if (char === "}" || char === "]") {
-            level--;
-            if (level === 0) {
-                sections.push([""]);
-            }
-            else {
-                sections[sections.length - 1][0] += char;
-            }
-        }
-        else if (char === "{" || char === "[") {
-            if (level !== 0) {
-                sections[sections.length - 1][0] += char;
-            }
-            else {
-                sections[sections.length - 1].push(char);
-            }
-            level++;
-        }
-        else {
-            sections[sections.length - 1][0] += char;
-        }
-    }
-    sections = sections.slice(0, sections.length - 1);
-    return sections;
-}
-
-function findFirst(string, subs, start) {
-    var min = -1;
-    var index;
-    var first = subs[0];
-    for (var i = 0; i < subs.length; i++) {
-        index = string.indexOf(subs[i], start);
-        if ((index !== -1) && (index < min || min === -1)) {
-            min = index;
-            first = subs[i];
-        }
-    }
-    return [min, first];
-}
-
-function parseNewCommand(definition, command, cmds) {
-    definition = definition.substring(command.length);
-    var sections = findBracketSections(definition);
-    if (sections.length === 0) {
-        var commands = definition.match(/\\[a-zA-Z]+/g);
-        cmds[commands[0]] = commands[1];
-    }
-    else if (sections.length === 1) {
-        var cmd = definition.replaceAll(sections[0][0], "").match(/\\[a-zA-Z]+/)[0];
-        if (definition.indexOf(cmd) === 0) {
-            cmds[cmd] = sections[0][0];
-        }
-        else {
-            cmds[sections[0][0]] = cmd;
-        }
-    }
-    else if (sections.length === 2) {
-        if (sections[0][1] !== "[") {
-            cmds[sections[0][0]] = sections[1][0];
-        }
-    }
-}
-
-function bracketBalance(string) {
-    var balance = 0;
-    var open = ["(", "[", "{"];
-    for (var index in open) {
-        balance += string.split(open[index]).length - 1;
-    }
-    var close = [")", "]", "}"];
-    for (index in close) {
-        balance -= string.split(close[index]).length - 1;
-    }
-    return balance;
-}
-
-function parseSty(inp, cmds) {
-    var starts = ["\\newcommand", "\\renewcommand", "\\DeclareRobustCommand", "\\defSpecFun"];
-    inp = inp.split("\n").map(function(line) {return line.match(/([^\\%]|\\.|^)+/g)[0];}).join("\n");
-    inp = inp.replace(/\n+/g, "\n").replace(/^\n|\n^/g, "");
-    var lines = inp.split("\n");
-    var lineno = 0;
-    while (lineno < lines.length) {
-        var line = lines[lineno];
-        var command = findFirst(line, starts, 0);
-        var balance = bracketBalance(line);
-        while (balance !== 0 || (lineno + 1 < lines.length && lines[lineno + 1].charAt(0) !== "\\")) {
-            lineno++;
-            balance += bracketBalance(lines[lineno]);
-            line += lines[lineno];
-        }
-        line = line.replace(/\s/g, "");
-        if (command[0] !== -1) {
-            command = command[1];
-            if (command === "\\newcommand" || command === "\\renewcommand") {
-                parseNewCommand(line, command, cmds);
-            }
-            else if (command === "\\DeclareRobustCommand") {
-                var sections = findBracketSections(line);
-                if (sections[0][0] !== undefined && sections[1][0] !== undefined) {
-                    cmds[sections[0][0]] = sections[1][0];
-                }
-            }
-        }
-        lineno++;
-    }
-}
-
 module.exports.paren = arr2set(["\\sinh"]);
 module.exports.other_literals3 = obj2map({});
-try{ //try catch to make sure program still runs without csv file
-    var csv = require('csv-parse/lib/sync');
-    var fs = require('fs');
-    var path = require('path');
-    module.exports.other_literals3 = {};
-    parseSty(fs.readFileSync(path.join(__dirname, "texvc.sty"), 'utf8'), module.exports.other_literals3);
+var csv = require('csv-parse/lib/sync');
+var fs = require('fs');
+var path = require('path');
+try {
+    module.exports.other_literals3 = parseSty(fs.readFileSync(path.join(__dirname, "texvc.sty"), 'utf8'));
+}
+catch (e) {
+    if (e.code !== "ENOENT") {
+        throw e;
+    }
+}
+try {
     var filePathread = path.join(__dirname, 'optionalFunctions.csv');
     var paren = [];
     var inp = fs.readFileSync(filePathread, 'utf8');
@@ -176,7 +70,12 @@ try{ //try catch to make sure program still runs without csv file
         c++;
     }
     module.exports.paren = arr2set(paren);
-} catch (e) {}
+}
+catch (e) {
+    if (e.code !== "ENOENT") {
+        throw e;
+    }
+}
 
 module.exports.mediawiki_function_names = arr2set([
     "\\arccot", "\\arcsec", "\\arccsc", "\\sgn", "\\sen"
@@ -633,82 +532,6 @@ module.exports.other_literals2 = arr2set([
     "\\textvisiblespace",
     "\\varstigma"
 ]);
-
-/*module.exports.other_literals3 = obj2map({
-    "\\C": "\\mathbb{C}",
-    "\\H": "\\mathbb{H}",
-    "\\N": "\\mathbb{N}",
-    "\\Q": "\\mathbb{Q}",
-    "\\R": "\\mathbb{R}",
-    "\\Z": "\\mathbb{Z}",
-    "\\alef": "\\aleph",
-    "\\alefsym": "\\aleph",
-    "\\Alpha": "\\mathrm{A}",
-    "\\and": "\\land",
-    "\\ang": "\\angle",
-    "\\Beta": "\\mathrm{B}",
-    "\\bull": "\\bullet",
-    "\\Chi": "\\mathrm{X}",
-    "\\clubs": "\\clubsuit",
-    "\\cnums": "\\mathbb{C}",
-    "\\Complex": "\\mathbb{C}",
-    "\\Dagger": "\\ddagger",
-    "\\diamonds": "\\diamondsuit",
-    "\\Doteq": "\\doteqdot",
-    "\\doublecap": "\\Cap",
-    "\\doublecup": "\\Cup",
-    "\\empty": "\\emptyset",
-    "\\Epsilon": "\\mathrm{E}",
-    "\\Eta": "\\mathrm{H}",
-    "\\exist": "\\exists",
-    "\\ge": "\\geq",
-    "\\gggtr": "\\ggg",
-    "\\hAar": "\\Leftrightarrow",
-    "\\harr": "\\leftrightarrow",
-    "\\Harr": "\\Leftrightarrow",
-    "\\hearts": "\\heartsuit",
-    "\\image": "\\Im",
-    "\\infin": "\\infty",
-    "\\Iota": "\\mathrm{I}",
-    "\\isin": "\\in",
-    "\\Kappa": "\\mathrm{K}",
-    "\\larr": "\\leftarrow",
-    "\\Larr": "\\Leftarrow",
-    "\\lArr": "\\Leftarrow",
-    "\\le": "\\leq",
-    "\\lrarr": "\\leftrightarrow",
-    "\\Lrarr": "\\Leftrightarrow",
-    "\\lrArr": "\\Leftrightarrow",
-    "\\Mu": "\\mathrm{M}",
-    "\\natnums": "\\mathbb{N}",
-    "\\ne": "\\neq",
-    "\\Nu": "\\mathrm{N}",
-    "\\O": "\\emptyset",
-    "\\omicron": "\\mathrm{o}",
-    "\\Omicron": "\\mathrm{O}",
-    "\\or": "\\lor",
-    "\\part": "\\partial",
-    "\\plusmn": "\\pm",
-    "\\rarr": "\\rightarrow",
-    "\\Rarr": "\\Rightarrow",
-    "\\rArr": "\\Rightarrow",
-    "\\real": "\\Re",
-    "\\reals": "\\mathbb{R}",
-    "\\Reals": "\\mathbb{R}",
-    "\\restriction": "\\upharpoonright",
-    "\\Rho": "\\mathrm{P}",
-    "\\sdot": "\\cdot",
-    "\\sect": "\\S",
-    "\\spades": "\\spadesuit",
-    "\\sub": "\\subset",
-    "\\sube": "\\subseteq",
-    "\\supe": "\\supseteq",
-    "\\Tau": "\\mathrm{T}",
-    "\\thetasym": "\\vartheta",
-    "\\varcoppa": "\\mbox{\\coppa}",
-    "\\weierp": "\\wp",
-    "\\Zeta": "\\mathrm{Z}"
-});*/
 
 module.exports.big_literals = arr2set([
     "\\big",
