+++ test/parseSty.sty
@@ -1,0 +1,8 @@
+\ProvidesPackage{parseSty}
+
+\newcommand\testA{\TeA}
+\newcommand{\testB}\TeB
+\newcommand{\testC}{\TeC}
+\newcommand{\testD}{\mbox{\TeD}}
+
+\endinput
